[website] merge token docs from dev into design
