+++ website/docs/style/design-tokens/components/design-tokens.module.css
@@ -3,8 +3,21 @@
   max-width: 400px;
 }
 
+.tokenNameWrapper {
+  max-width: 100%;
+}
 .tokenName {
   cursor: pointer;
+  max-width: 100%;
+}
+
+.tokenValueWrapper {
+  max-width: 100%;
+}
+.tokenValue {
+  cursor: pointer;
+  display: flex;
+  max-width: 100%;
 }
 
 .colorPreview {
@@ -18,5 +31,6 @@
 }
 
 .usages {
-  text-decoration: underline;
+  color: #6c6e79;
+  border-bottom: 1px dashed #6c6e79;
 }
