+++ website/docs/style/design-tokens/components/design-tokens.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './design-tokens.module.css';
 import Input from '@semcore/input';
 import SearchIcon from '@semcore/icon/search/m';
-import designTokens from './tokens-list.json';
+import designTokens from './design-tokens.json';
 import DataTable from '@semcore/data-table';
 import Link from '@semcore/link';
 import Tooltip from '@semcore/tooltip';
@@ -10,7 +10,7 @@ import Ellipsis, { useResizeObserver } from '@semcore/ellipsis';
 import { brightness } from '@semcore/utils/lib/color';
 import Copy from '@components/Copy';
 
-const ColorPreview: React.FC<{ color: string }> = ({ color }) => {
+export const ColorPreview: React.FC<{ color: string }> = ({ color }) => {
   if (!color.startsWith('#') && !color.startsWith('rgba(')) return null;
 
   const needsBorder = React.useMemo(
@@ -33,16 +33,18 @@ const DesignTokens: React.FC = () => {
   const [filter, setFilter] = React.useState('');
   const filteredTokens = React.useMemo(
     () =>
-      designTokens.filter(({ name, defaultValue, description }) => {
+      designTokens.filter(({ name, rawValue, description }) => {
         if (!filter) return true;
         if (name?.toLowerCase().includes(filter.toLowerCase())) return true;
-        if (defaultValue?.toLowerCase().includes(filter.toLowerCase())) return true;
+        if (rawValue?.toLowerCase().includes(filter.toLowerCase())) return true;
         if (description?.toLowerCase().includes(filter.toLowerCase())) return true;
         return false;
       }),
     [filter],
   );
 
+  const nameHeaderRef = React.useRef(null);
+  const nameHeaderRect = useResizeObserver(nameHeaderRef);
   const valueHeaderRef = React.useRef(null);
   const valueHeaderRect = useResizeObserver(valueHeaderRef);
   const descriptionHeaderRef = React.useRef(null);
@@ -56,44 +58,61 @@ const DesignTokens: React.FC = () => {
       </Input>
       <DataTable data={filteredTokens}>
         <DataTable.Head>
-          <DataTable.Column name="name" children="Token name" wMin={350} flex="1 0 auto" />
-          <DataTable.Column name="defaultValue" children="Value" ref={valueHeaderRef} />
+          <DataTable.Column name="name" children="Token name" ref={nameHeaderRef} />
+          <DataTable.Column name="value" children="Value" ref={valueHeaderRef} />
           <DataTable.Column name="description" children="Description" ref={descriptionHeaderRef} />
           <DataTable.Column name="components" children="Used in" />
         </DataTable.Head>
-        <DataTable.Body virtualScroll={{ rowHeight: 45 }} h={800}>
+        <DataTable.Body virtualScroll={{ rowHeight: 45, tollerance: 10 }} h={800}>
           <DataTable.Cell name="name">
             {(props, row) => {
               return {
                 children: (
-                  <div>
-                    <Copy
-                      title="Copied"
-                      text={row[props.name]}
-                      trigger="click"
-                      className={styles.tokenName}
-                    >
-                      <div>{row[props.name]}</div>
-                    </Copy>
-                  </div>
+                  <Copy
+                    title="Copied"
+                    text={row[props.name]}
+                    textTooltip={`Click to copy "${row[props.name]}"`}
+                    trigger="click"
+                    className={styles.tokenNameWrapper}
+                  >
+                    <div className={styles.tokenName}>
+                      <Ellipsis
+                        trim="middle"
+                        tooltip={false}
+                        containerRect={nameHeaderRect}
+                        containerRef={nameHeaderRef}
+                      >
+                        {row[props.name]}
+                      </Ellipsis>
+                    </div>
+                  </Copy>
                 ),
               };
             }}
           </DataTable.Cell>
-          <DataTable.Cell name="defaultValue">
+          <DataTable.Cell name="value">
             {(props, row) => {
               return {
                 children: (
-                  <>
-                    <ColorPreview color={row[props.name]} />
-                    <Ellipsis
-                      trim="middle"
-                      containerRect={valueHeaderRect}
-                      containerRef={valueHeaderRef}
-                    >
-                      {row[props.name]}
-                    </Ellipsis>
-                  </>
+                  <Copy
+                    title="Copied"
+                    text={row.rawValue}
+                    textTooltip={`Click to copy "${row.rawValue}"`}
+                    trigger="click"
+                    className={styles.tokenValueWrapper}
+                  >
+                    <div className={styles.tokenValue}>
+                      <ColorPreview color={row.computedValue} />
+                      <Ellipsis
+                        trim="middle"
+                        tooltip={false}
+                        containerRect={valueHeaderRect}
+                        containerRef={valueHeaderRef}
+                      >
+                        {row.rawValue}
+                      </Ellipsis>
+                    </div>
+                  </Copy>
                 ),
               };
             }}
@@ -103,7 +122,7 @@ const DesignTokens: React.FC = () => {
               return {
                 children: (
                   <Ellipsis
-                    trim="end"
+                    trim="middle"
                     containerRect={descriptionHeaderRect}
                     containerRef={descriptionHeaderRef}
                   >
@@ -138,9 +157,7 @@ const DesignTokens: React.FC = () => {
                 children: (
                   <>
                     <Tooltip>
-                      <Tooltip.Trigger className={styles.usages}>
-                        {row[props.name].length} components
-                      </Tooltip.Trigger>
+                      <Tooltip.Trigger className={styles.usages}>hint links</Tooltip.Trigger>
                       <Tooltip.Popper>
                         {row[props.name].map((componentName, index) => (
                           <React.Fragment key={componentName}>
